Make excluding files from linting easier
By default all source files in the project are statically check for
style issues and errors with the in-house version of `cpplint`. If any
errors are encountered, the build fails.  At the same time, some of the
source files come from external projects and are included as-is and must
be excluded from the linting process as their formatting does not
conform with the project's requirements.  The list of these files was
hard-coded in to the CMake function responsible for linting.  I modified
this function to allow one to pass patterns which will be used to filter
out files for which the linting process should be skipped.